unicode: Merged from trunk up to [5308].
git-svn-id: http://code.djangoproject.com/svn/django/branches/unicode@5310 bcc190cf-cafb-0310-a4f2-bffc1f526a37

+++ django/db/backends/mysql/introspection.py
@@ -76,7 +76,7 @@ def get_indexes(cursor, table_name):
 DATA_TYPES_REVERSE = {
     FIELD_TYPE.BLOB: 'TextField',
     FIELD_TYPE.CHAR: 'CharField',
-    FIELD_TYPE.DECIMAL: 'FloatField',
+    FIELD_TYPE.DECIMAL: 'DecimalField',
     FIELD_TYPE.DATE: 'DateField',
     FIELD_TYPE.DATETIME: 'DateTimeField',
     FIELD_TYPE.DOUBLE: 'FloatField',
